requirements.txt
streamlit>=1.35
pandas
numpy
altair<6
matplotlib
plotly
scipy
scikit-learn
requests
python-dotenv
